token aware, refactor
