+++ src/screens/Ticketing/Tickets/AvailableTickets/AvailableTickets.tsx
@@ -14,6 +14,8 @@ import {
   TransportSubmode,
 } from '@atb/api/types/generated/journey_planner_v3_types';
 import {useFirestoreConfiguration} from '@atb/configuration/FirestoreConfigurationContext';
+import {productIsSellableInApp} from '@atb/reference-data/utils';
+import Ticket from '@atb/screens/Ticketing/Tickets/AvailableTickets/Ticket';
 
 export const AvailableTickets = ({
   onBuySingleTicket,
@@ -23,21 +25,29 @@ export const AvailableTickets = ({
   onBuyPeriodTicket: () => void;
 }) => {
   const styles = useStyles();
-  const {theme} = useTheme();
-  const {abtCustomerId, authenticationType} = useAuthState();
   const hasEnabledMobileToken = useHasEnabledMobileToken();
   const {preassignedFareproducts} = useFirestoreConfiguration();
   const {t} = useTranslation();
 
-  const isSignedInAsAbtCustomer = !!abtCustomerId;
+  const shouldShowSingleTicket = preassignedFareproducts
+    .filter(productIsSellableInApp)
+    .some((product) => {
+      return product.type === 'single';
+    });
 
-  // TODO check firestore for availability
+  const shouldShowPeriodTicket =
+    hasEnabledMobileToken &&
+    preassignedFareproducts.filter(productIsSellableInApp).some((product) => {
+      return product.type === 'period';
+    });
+
+  const shouldShowSummerPass = false;
 
   return (
     <View style={styles.container}>
-      {isSignedInAsAbtCustomer && (
-        <ScrollView>
-          <View style={styles.ticketsContainer}>
+      <ScrollView>
+        <View style={styles.ticketsContainer}>
+          {shouldShowSingleTicket && (
             <Ticket
               title={t(TicketsTexts.availableTickets.singleTicket.title)}
               transportationModeTexts={t(
@@ -52,6 +62,8 @@ export const AvailableTickets = ({
               icon={TicketIcons.Single}
               onPress={onBuySingleTicket}
             />
+          )}
+          {shouldShowPeriodTicket && (
             <Ticket
               title={t(TicketsTexts.availableTickets.periodTicket.title)}
               transportationModeTexts={t(
@@ -66,25 +78,9 @@ export const AvailableTickets = ({
               icon={TicketIcons.Period}
               onPress={onBuyPeriodTicket}
             />
-          </View>
-          <View style={styles.ticketsContainer}>
-            <Ticket
-              title={t(TicketsTexts.availableTickets.summerPass.title)}
-              transportationModeTexts={t(
-                TicketsTexts.availableTickets.summerPass.transportModes,
-              )}
-              transportationModeIcons={[
-                {mode: Mode.Bus, subMode: TransportSubmode.LocalBus},
-                {mode: Mode.Rail},
-                {mode: Mode.Water},
-              ]}
-              description={t(
-                TicketsTexts.availableTickets.summerPass.description,
-              )}
-              icon={TicketIcons.Summer}
-              onPress={onBuyPeriodTicket}
-            />
-          </View>
+          )}
+        </View>
+        {shouldShowSummerPass && (
           <View style={styles.ticketsContainer}>
             <Ticket
               title={t(TicketsTexts.availableTickets.summerPass.title)}
@@ -104,76 +100,8 @@ export const AvailableTickets = ({
               onPress={onBuyPeriodTicket}
             />
           </View>
-        </ScrollView>
-      )}
-    </View>
-  );
-};
-
-type TransportationModeIconProperties = {
-  mode: Mode;
-  subMode?: TransportSubmode;
-};
-
-const Ticket = ({
-  title,
-  description,
-  transportationModeIcons,
-  transportationModeTexts,
-  icon,
-  accented = false,
-  onPress,
-}: {
-  title: string;
-  description: string;
-  transportationModeIcons: TransportationModeIconProperties[];
-  transportationModeTexts: string;
-  icon: (props: SvgProps) => JSX.Element;
-  accented?: boolean;
-  onPress: () => void;
-}) => {
-  const styles = useStyles();
-  const ticketTheme = accented ? styles.ticket_accented : styles.ticket_normal;
-  const textColor = accented ? 'primary_2' : 'primary';
-
-  return (
-    <View style={[styles.ticket, ticketTheme]}>
-      <TouchableOpacity onPress={onPress}>
-        <View style={{flexShrink: 1}}>
-          <View style={{flex: 1, flexDirection: 'row', alignItems: 'center'}}>
-            {transportationModeIcons.map((icon) => {
-              return (
-                <TransportationIcon mode={icon.mode} subMode={icon.subMode} />
-              );
-            })}
-            <ThemeText
-              type="body__tertiary"
-              style={styles.transportation_label}
-              color={textColor}
-            >
-              {transportationModeTexts}
-            </ThemeText>
-          </View>
-          <ThemeText
-            type="body__secondary--bold"
-            style={styles.ticket_name}
-            accessibilityLabel={title}
-            color={textColor}
-          >
-            {title}
-          </ThemeText>
-          <ThemeText
-            type="body__tertiary"
-            style={styles.description}
-            color={textColor}
-          >
-            {description}
-          </ThemeText>
-        </View>
-        <View style={styles.ticketIllustrationContainer}>
-          <View style={styles.ticketIllustration}>{icon({})}</View>
-        </View>
-      </TouchableOpacity>
+        )}
+      </ScrollView>
     </View>
   );
 };
@@ -183,9 +111,6 @@ const useStyles = StyleSheet.createThemeHook((theme) => ({
     flex: 1,
     backgroundColor: theme.colors.background_1.backgroundColor,
   },
-  buyPeriodTicketButton: {
-    marginTop: theme.spacings.medium,
-  },
 
   ticketsContainer: {
     flex: 1,
@@ -194,39 +119,4 @@ const useStyles = StyleSheet.createThemeHook((theme) => ({
     paddingBottom: theme.spacings.medium,
     alignItems: 'stretch',
   },
-  ticket: {
-    width: '100%',
-    flexShrink: 1,
-    alignSelf: 'stretch',
-    marginRight: theme.spacings.medium,
-    padding: theme.spacings.xLarge,
-    borderRadius: theme.border.radius.regular,
-  },
-
-  ticket_normal: {
-    backgroundColor: theme.colors.background_0.backgroundColor,
-  },
-  ticket_accented: {
-    backgroundColor: theme.colors.primary_2.backgroundColor,
-    textColor: theme.colors.primary_2.color,
-  },
-
-  ticketIllustrationContainer: {
-    flexGrow: 1,
-    flexDirection: 'row',
-    marginTop: theme.spacings.small,
-  },
-  ticketIllustration: {
-    alignSelf: 'flex-end',
-    opacity: 0.6,
-  },
-  label_uppercase: {
-    // TODO: this will be paert of design system
-  },
-  transportation_label: {},
-  ticket_name: {
-    marginBottom: theme.spacings.small,
-    marginTop: theme.spacings.small,
-  },
-  description: {},
 }));
